Verify bind config with named-checkconf
