+++ app/lib/use_cases/result.rb
@@ -1,0 +1,13 @@
+module UseCases
+  class Result
+    attr_reader :error
+
+    def initialize(error = nil)
+      @error = error
+    end
+
+    def success?
+      error.nil?
+    end
+  end
+end
