+++ app/lib/use_cases/verify_bind_config.rb
@@ -7,24 +7,12 @@ module UseCases
     def call(config)
       bind_verifier_gateway.verify_config(config)
       Result.new
-    rescue Gateways::BindVerifier::InternalError => error
+    rescue Gateways::BindVerifier::ConfigurationError => error
       Result.new(error)
     end
 
     private
 
     attr_reader :bind_verifier_gateway
-
-    class Result
-      attr_reader :error
-
-      def initialize(error = nil)
-        @error = error
-      end
-
-      def success?
-        error.nil?
-      end
-    end
   end
 end
